fix(maps): Ensure previous map pins are cleared when rendering
When we have a query result, we render pins to the map, when we
then do a subsequent query, we do not clear the previous map data
even though we do clear the context data.

Keep a reference array of our map markers when we create them,
then clear out the referemce array every time we run ondata.